ml/requirements.txt
ultralytics>=8.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
shapely>=2.0.0
